页面重新布局 & 发货地

+++ TxoooProductUpload/TxoooProductUpload/Service/Entities/ProductResult.cs
@@ -74,7 +74,7 @@ namespace TxoooProductUpload.Service.Entities
         /// <summary>
         /// 商品类别id
         /// </summary>
-        public int product_type { set; get; }
+        public long product_type { set; get; }
 
         /// <summary>
         /// 商品详情
